Support Remote jobs in Folders
This change adds a generateJobPath() helper function which turns job
names into encoded job names. Job names are split on '/', each part is
run through encodeValue(), then the resulting encoded parts are joined
by "/job/". All instances where encodeValue(job) was used to build a job
URL have been replaced with calls to generateJobPath().

A new test was added to handle the case where a job is in a folder.